Created two trigger for cart.
Created two triggers for the cart, so the First trigger will check that TotalPrice should not be negative when a new row is inserted or updated in the cart table. And in the second trigger, it will back up data to a new table(cart_backup) when a row is deleted from the original cart table.